recommender_ai_service/requirements.txt
# ── Django ─────────────────────────────────────────────────────────────────
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-redis>=5.4

# ── Celery / Task queue ────────────────────────────────────────────────────
celery>=5.3
redis>=5.0

# ── Database ───────────────────────────────────────────────────────────────
psycopg2-binary>=2.9   # PostgreSQL (prod)

# ── PyTorch (CPU build — change to torch[cuda] for GPU) ───────────────────
torch>=2.0
numpy>=1.26

# ── Data / ETL ─────────────────────────────────────────────────────────────
pandas>=2.1
scikit-learn>=1.4
pyarrow>=14.0           # parquet support

# ── Network / HTTP ─────────────────────────────────────────────────────────
requests>=2.31

# ── Server ─────────────────────────────────────────────────────────────────
gunicorn>=21.2
uvicorn[standard]>=0.27

# ── Env / Config ───────────────────────────────────────────────────────────
python-dotenv>=1.0

# ── Testing ────────────────────────────────────────────────────────────────
pytest>=7.4
pytest-django>=4.7
